Started working on code generation on-the-fly
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@520 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/wxSmith/properties/wxsstringlistproperty.cpp
@@ -61,7 +61,7 @@ namespace {
 			
 				CenterOnScreen();
 				
-				for ( int i=0; i<Array.Count(); i++ )
+				for ( int i=0; i<(int)Array.Count(); i++ )
 				{
 					List->AppendText(Array[i]);
 					List->AppendText(wxT("\n"));
